+++ README.md
@@ -10,6 +10,7 @@ are extensions of [Fluent UI](https://github.com/microsoft/fluentui).
 | Vertical Virtualized DataGrid (react-window)            | [README.md](https://github.com/microsoft/fluentui-contrib/blob/main/packages/react-data-grid-react-window/README.md)      | [Storybook](https://microsoft.github.io/fluentui-contrib/react-data-grid-react-window/?path=/story/datagrid--virtualized-data-grid)      | [![npm version](https://img.shields.io/npm/v/@fluentui-contrib/react-data-grid-react-window?style=flat-square)](https://www.npmjs.com/package/@fluentui-contrib/react-data-grid-react-window)      |
 | 2d Virtualized DataGrid (react-window)                  | [README.md](https://github.com/microsoft/fluentui-contrib/blob/main/packages/react-data-grid-react-window-grid/README.md) | [Storybook](https://microsoft.github.io/fluentui-contrib/react-data-grid-react-window-grid/?path=/story/datagrid--virtualized-data-grid) | [![npm version](https://img.shields.io/npm/v/@fluentui-contrib/react-data-grid-react-window?style=flat-square)](https://www.npmjs.com/package/@fluentui-contrib/react-data-grid-react-window-grid) |
 | Resize Handle for dynamic layouts (react-resize-handle) | [README.md](https://github.com/microsoft/fluentui-contrib/blob/main/packages/react-resize-handle/README.md)               | [Storybook](https://microsoft.github.io/fluentui-contrib/react-resize-handle)                                                            | [![npm version](https://img.shields.io/npm/v/@fluentui-contrib/react-resize-handle?style=flat-square)](https://www.npmjs.com/package/@fluentui-contrib/react-resize-handle)                        |
+| react-tree-grid                                         | [README.md](https://github.com/microsoft/fluentui-contrib/blob/main/packages/react-tree-grid/README.md)                   | [Storybook](https://microsoft.github.io/fluentui-contrib/react-tree-grid)                                                                | [![npm version](https://img.shields.io/npm/v/@fluentui-contrib/react-tree-grid?style=flat-square)](https://www.npmjs.com/package/@fluentui-contrib/react-tree-grid)                                |
 
 ## [Contributing](./Contributing.md)
