+++ change/@fluentui-contrib-react-tree-grid-9e028783-109e-40ba-9687-19339afa5836.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "chore: release",
+  "packageName": "@fluentui-contrib/react-tree-grid",
+  "email": "bernardo.sunderhus@gmail.com",
+  "dependentChangeType": "patch"
+}
