+++ packages/react-tree-grid/src/components/TreeGridRow/TreeGridRow.types.ts
@@ -1,14 +1,17 @@
-import { Slot } from '@fluentui/react-components';
+import { ComponentProps, Slot } from '@fluentui/react-components';
 import { EventData, EventHandler } from '@fluentui/react-utilities';
 import * as React from 'react';
-import { ComponentProps } from '../../utils/types';
+
+export type TreeGridRowSlots = {
+  root: Slot<'div'>;
+};
 
 export type TreeGridRowOnOpenChangeData = { open: boolean } & (
   | EventData<'click', React.MouseEvent<HTMLDivElement>>
   | EventData<'keydown', React.KeyboardEvent<HTMLDivElement>>
 );
 
-export type TreeGridRowProps = ComponentProps<Slot<'div'>> & {
+export type TreeGridRowProps = ComponentProps<TreeGridRowSlots> & {
   open?: boolean;
   defaultOpen?: boolean;
   onOpenChange?: EventHandler<TreeGridRowOnOpenChangeData>;
