+++ packages/react-tree-grid/src/index.ts
@@ -1,14 +1,18 @@
 export { TreeGrid } from './components/TreeGrid';
-export type { TreeGridProps } from './components/TreeGrid';
+export type { TreeGridProps, TreeGridSlots } from './components/TreeGrid';
 
 export { TreeGridRow } from './components/TreeGridRow';
 export type {
   TreeGridRowProps,
   TreeGridRowOnOpenChangeData,
+  TreeGridRowSlots,
 } from './components/TreeGridRow';
 
 export { TreeGridCell } from './components/TreeGridCell';
-export type { TreeGridCellProps } from './components/TreeGridCell';
+export type {
+  TreeGridCellProps,
+  TreeGridCellSlots,
+} from './components/TreeGridCell';
 
 export {
   TreeGridRowProvider,
@@ -17,4 +21,7 @@ export {
 export type { TreeGridRowContextValue } from './contexts/TreeGridRowContext';
 
 export { TreeGridInteraction } from './components/TreeGridInteraction';
-export type { TreeGridInteractionProps } from './components/TreeGridInteraction';
+export type {
+  TreeGridInteractionProps,
+  TreeGridInteractionSlots,
+} from './components/TreeGridInteraction';
