--- packages/react-tree-grid/src/utils/types.ts
@@ -1,14 +1,0 @@
-import {
-  DistributiveOmit,
-  ExtractSlotProps,
-  SlotShorthandValue,
-  UnknownSlotProps,
-} from '@fluentui/react-utilities';
-
-export type ComponentProps<
-  S extends UnknownSlotProps | SlotShorthandValue | null | undefined
-> = PropsWithoutRef<ExtractSlotProps<S>>;
-
-type PropsWithoutRef<P> = 'ref' extends keyof P
-  ? DistributiveOmit<P, 'ref'>
-  : P;
